pyproject.toml
[tool.poetry]
name = "playwright-test"
version = "0.1.0"
description = ""
authors = ["tomoyachen <tomoya_chen@163.com>"]

[tool.poetry.dependencies]
python = "^3.8"
playwright = "^1.10.0"
pytest-playwright = "^0.1.0"
pytest-html = "^3.1.1"
pytest-rerunfailures = "^9.1.1"
pyyaml = "^5.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
